[UPDATED] Ocean Shader Complete!-ish
I think this is as far I'm going to go with this water shader. I cleaned it up and clamped a bunch of values to it can be usable.
Next step would be to add it to the main scene and merge this into the android dev branch and see how it preforms on mobile. and prepare for optimization hell. ;-;